Fix the loading bug but not completely
There was a bug that app is stopped when load the app. It wasn't there during debugging.

I exported to an apk file and then the bug occured.

I was trying to figure out what the reason for the bug. and I did below things.

But It still occurs sometimes when load the app. but not frequently.

- Removed theme on the loading component.
  then app'd not stopped while loading component.

- Added permissions on android in app.json.
  This needs for publishing to the google play store.

- Chnaged StyleSheet Object to JSON in PlanList.
  then app'd not stopped after loading.

but app's stopped after loading sometimes. Approximately It oucurs once in 20 times.

- Reasons for the bug that I think

  - Use StyleSheet intead of JSON. because of dinamical theme.

  - Problems of React navigation module. because I had some troubles by that.

  - Problems while fetching persist data with Async in redux.

  - Problems while updating from expo server like network problems.

I ended this issue. I'll move to another project.
and at next project, I'll consider testing more during developing.